Fix slurmstepd unintentionally killing itself from _sig_agent() thread
proctrack_g_get_pids() can return not just the stepd task pids, but also the
stepd pid itself depending on configuration. With proctrack/cgroup (v2),
the hook implementation also returns the stepd pid.

The pid check could be compared with the cont_id, but the cont_id only
represents the stepd pid if proctrack/cgroup or linuxproc is configured.
If pgid is configured, then cont_id represents the forked tasks PGID
(pid of the first forked task). Thus, relying on getpid() instead of
cont_id for the check to be configuration agnostic.

Changelog: Fix slurmstepd unintentionally killing itself if proctrack/cgroup
 and cgroup/v2 configured while deferring killing tasks due to any still
 core dumping.
Ticket: 22593
Cherry-picked: 5e5bb473c6

+++ src/interfaces/proctrack.c
@@ -288,6 +288,7 @@ static void *_sig_agent(void *args)
 {
 	bool hung_pids = false;
 	sig_agent_arg_t *agent_arg_ptr = args;
+	pid_t stepd_pid = getpid();
 
 	while (1) {
 		pid_t *pids = NULL;
@@ -332,6 +333,9 @@ static void *_sig_agent(void *args)
 			}
 
 			for (i = 0; i < npids; i++) {
+				/* Avoid killing our own (stepd) process. */
+				if (pids[i] == stepd_pid)
+					continue;
 				/* Kill processes */
 				kill(pids[i], agent_arg_ptr->signal);
 			}
